[WIP] core, ldelf: run TA initialization functions (.init_array)
Add support for executing functions referenced by the DT_INIT_ARRAY
entry of the dynamic segment of any ELF file loaded by a TA (main
executable or library). DT_INIT_ARRAY normally points to the
.init_array section, which is an array of function pointers that should
be called in sequence after the binary is mapped and before the entry
point is called.

Initialization functions can be declared in C as follows:

 void __attribute__((constructor)) foo(void)
 {
    ...
 }

TBD: Actually call the fn pointers
TBD: Trigger the same mechanism on dlopen()

Signed-off-by: Jerome Forissier <jerome@forissier.org>