Disable alignment check by default
Preparing for C++ support in TAs.

CFG_SCTLR_ALIGNMENT_CHECK is 'y' by default which prevents running TAs
that are not built with the strict alignment compiler options. If the
TA happens to be using a third party library, it can be problematic. An
example is libgcc_eh.a, the C++ exception handling library.

I see no reason to enforce alignment by default. Therefore, set
CFG_SCTLR_ALIGNMENT_CHECK to 'n'.

Signed-off-by: Jerome Forissier <jerome@forissier.org>