ci: duplicate Shippable tests on CodeShip
Shippable has a maximum build time of 1 hour for Open Source projects.
It is usually sufficient since typical build time is currently around
35 minutes, thanks to caching (ccache). But it can happen that the code
changes so much in a PR that ccache is ineffective and build time would
exceed 1 hour. When a build timeout occurs, the cache is NOT updated
and there is no other option that to manually feed the cache by forcing
a partial build (remove some steps in .shippable.yml). This is quite
annoying.

This commit introduces the required files to run the CI on CodeShip Pro
[1]. Its use is free for Open Source projects [2] and the maximum build
time is 5 hours [3] which should leaves use quite some room for the
occasional long builds.

The ccache directory is currently limited to 1GB but could be bigger if
needed. The compressed archive is uploaded to and downloaded from a
private server using scp. The private key is stored in an encrypted
environment variable: ID_RSA in codeship-env.encrypted. The encryption
key is private to the project and can be obtained by project admins
from codeship.com. CodeShip provides a tool called 'jet' to encrypt a
file (it reads the key from local file codeship.aes):

  jet encrypt codeship-env codeship-env.encrypted

Link: [1] https://codeship.com/features/pro
Link: [2] https://codeship.com/pricing/pro#:~:text=Open%20source%20projects%20will%20always%20be%20free
Link: [3] https://documentation.codeship.com/pro/builds-and-configuration/steps/#step-timeouts
Signed-off-by: Jerome Forissier <jerome@forissier.org>