mk/lib.mk: cleanup shared library link command
The command used to link shared libraries when CFG_ULIBS_SHARED=y is
slightly incorrect for two reasons:

1. The -L/-l arguments are passed before the object files, when they
should normally be added after;

2. The shared libraries needed during the link are passed as files in
addition to being supplied with -L/-l. This is redundant, and is a
consequence of having the shared libraries in the prerequisites and
using $^. Order-only prerequisites should be used instead.

Signed-off-by: Jerome Forissier <jerome@forissier.org>