ci: shippable: add secondary caching
Shippable has a maximum build time of 1 hour for Open Source projects.
It is usually sufficient since typical build time is currently around
35 minutes, thanks to caching (ccache). But it can happen that the code
changes too much in a PR so ccache is ineffective and build time would
exceed 1 hour. When a build timeout occurs, the cache is NOT updated
and there is no other option than to manually feed the cache by forcing
a partial build (remove some steps in .shippable.yml). This is quite
impractical.

This commit introduces a secondary cache on a private server with an
upload in the middle of the build. This way, if the build times out we
are still likely to have cached a significant amount of data and
restarting the build should be enough to address the issue. If the
remote server is unavailable for any reason, the errors are ignored and
the build occurs as before with the cache stored in the Shippable
infrastructure.

The compressed archive is uploaded and downloaded using scp. Each
project on Shipable has its own SSH key, the public key just needs to
be configured on the cache server for this to work.

Signed-off-by: Jerome Forissier <jerome@forissier.org>