ci: shippable: add secondary cache
Shippable has a maximum build time of 1 hour for Open Source projects.
It is usually sufficient since typical build time is currently around
35 minutes, thanks to caching (ccache). But it can happen that the code
changes too much in a PR so ccache is ineffective and build time would
exceed 1 hour. When a build timeout occurs, the cache is NOT updated
and there is no other option than to manually feed the cache with a
partial build (by removing some steps in .shippable.yml) so that the
next build can be a bit quicker, wait for the build to succeed, then
restore all the build steps. This is quite impractical.

This commit introduces a secondary cache on a private server with an
upload in the middle of the build. This way, if the build times out we
are still likely to have cached a significant amount of data and
restarting the build should be enough to deal with the problem -- no
need to trim .shippable.yml. If the server is unavailable for any
reason, errors are ignored and the build runs as before using the
cache stored in the Shippable infrastructure.

The cache data are transferred using SSH. Since each Shippable project
has its own SSH keypair, the public key just needs to be configured on
the cache server for this to work with forks for instance.

Signed-off-by: Jerome Forissier <jerome@forissier.org>