ldelf, libutee: rework support of DT_INIT_ARRAY/DT_FINI_ARRAY
Now that we have the standard function dl_iterate_phdr() in libutee, we
can use it to process the initialization and finalization arrays in the
ELF files. By doing so, we can remove the ad-hoc structures introduced
in commit dd655cb9906c ("ldelf, ta: add support for DT_INIT_ARRAY and
DT_FINI_ARRAY"). This reduces the interface between ldelf and the TA.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>