libutils: move COMPILE_TIME_ASSERT() from <assert.h> to <assert_ext.h>
COMPILE_TIME_ASSERT() is not defined by the C language so it does not
belong in the standard header <assert.h>. Let's move it to a new file:
<assert_ext.h>, similar to other extensions we have added to libutils
(<stdlib_ext.h>, <string_ext.h> and <types_ext.h>).

The files that use this macro are modified to reflect the change. Either
<assert.h> is replaced with <assert_ext.h> (if assert() is not used) or
<assert_ext.h> is added (if assert() is used).

Signed-off-by: Jerome Forissier <jerome@forissier.org>