Discard section .note.gnu.property from ldelf and kernel ld.S
It is observed that clang compiler sometimes places the
.note.gnu.property at offset 0. This causes issues during
relocation as _ldelf_start() which is the entry point of the ldelf
should be first in the binary. Since this section is not parsed from
either tee.elf or ldelf.elf, discard it in linker to avoid such
conflicts.

This is done in preparation for BTI support in later commits.

Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>
Reviewed-by: Jerome Forissier <jerome@forissier.org>