scripts: add print_tee_hash.py
Adds a Python script to return the SHA256 hash of the .text and .rodata
sections of tee.elf, excluding the area in that contain relocatable data
or data initialized early after boot. This hash value can later be
compared with the hash returned at runtime by the  attestation PTA which
will come in a later commit.

Signed-off-by: Jerome Forissier <jerome@forissier.org>