core: add delimited area in .text to store data
A few variables such as boot_mmu_config are stored within the .text
section of tee.elf, because they need to be reachable from the identity
mapping which covers a subset of .text. Having them here however is a
problem when one wants to measure (hash) the .text section because the
runtime content may be different from the content in the tee.elf. In
order to workaround this issue, allocate an area at the end of the .text
section to gather the data that are modified at boot time. Symbols
tagges with .identity_map.data will be stored there. Two delimiters are
introduced: __text_data_start and __text_data_end.

Signed-off-by: Jerome Forissier <jerome@forissier.org>