core: do not tag ops structures with __rodata_unpaged
The global data structures currently tagged with __rodata_unpaged
contain relocatable data (function pointers) so they should reside
in a writeable section when ASLR is enabled. Therefore remove the
__rodata_unpaged attribute. With this change:
- With ASLR enabled (-fpic flag present), the compiler puts the
structures into .data which is entirely unpaged so there is no issue
pager enabled or not;
- With ASLR disabled, (-fpic not present), the compiler would put the
structures into .rodata. Therefore add DECLARE_KEEP_PAGER() so that the
data ends up in the unpaged part if the RO data like before this patch.
The end goal is to make .rodata truly read-only to enable remote
attestation.

Signed-off-by: Jerome Forissier <jerome@forissier.org>