core: bignum: mpi: allow NULL parameter to crypto_bignum_clear()
Allow to pass a NULL argument to crypto_bignum_clear() for convenience.
With this change it will become easier to clear a structure with big
numbers that can potentially be absent (NULL).

Note that mpi_desc.c is under core/lib/libtomcrypt but actually uses
MBedTLS big numbera code and is used both with LibTomCrypt and MBedTLS
as a core crypto library.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>