core: libmbedtls: set CRT parameters to NULL with MBEDTLS_RSA_NO_CRT
When using MBedTLS as the core crypto library, TEE_GenerateKey() always
allocates big numbers for the CRT parameters, but sets them to zero
because CRT is always disabled via MBEDTLS_RSA_NO_CRT [1]. In other
words, the CRT attributes are present in the key object but always zero.
The GP specification mentions that the CRT attributes are optional when
used in TEE_PopulateTransientObject(), but they are not in
TEE_GenerateKey(). This situation is a bit silly however: allocating
useless objects and claiming attributes are present when they are
actually not is not optimal.

Therefore, do not allocate CRT parameters nor set the corresponding
attribute bits when generating an RSA key pair with MBedTLS.

[1] MBEDTLS_RSA_NO_CRT is always enabled because the contrary would
mean CRT parameters are mandatory. That would not comply with the
specification of TEE_PopulateTransientObject() which defines the CRT
parameter as optional. A consequence of MBEDTLS_RSA_NO_CRT is that
TEE_GenerateKey() does not produce the CRT parameters either. Note
that LibTomCrypt has a different approach: it always produces the CRT
parameters when generating a key pair, but doesn't require them to work
with a key. Much like the GP specification.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>