core: crypto: do not set CRT attribute bits when not generated
When using MBedTLS as the core crypto library, TEE_GenerateKey() always
sets the CRT attributes to zero due to MBEDTLS_RSA_NO_CRT [1]. Although
the GP specification does not define them as optional in
TEE_GenerateKey(), they are optional in TEE_PopulateTransientObject()
and it makes more sense to not expose them to the TA in
TEE_GenerateKey() either if they have not been computed.

[1] MBEDTLS_RSA_NO_CRT is always enabled because the contrary would
mean CRT parameters are always mandatory. That would not comply with
the specification of TEE_PopulateTransientObject() which defines them
as optional. A consequence of MBEDTLS_RSA_NO_CRT is that
TEE_GenerateKey() does not produce the CRT parameters either. Note
that LibTomCrypt has a different approach: it always produces the CRT
parameters when generating a key pair, but doesn't require them to work
with a key. Much like the GP specification.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>