libmbedtls: bignum: restore mbedtls_mpi_exp_mod() from v3.5.2
The implementation of mbedtls_mpi_exp_mod() in Mbed TLS v3.6.0
introduces a large performance regression in "xtest 4011" on QEMUv7
(32-bit). One iteration of the test used to take 1.4 second on my
machine but the newer implementation now needs 23 seconds. To make
matters worse, xtest 4011 runs ten iterations so in reality it is
14 seconds vs. almost 4 minutes for the whole test.
Revert mbedtls_mpi_exp_mod() to the v3.5.2 implementation to gain the
performance back. The upstream commit that changed the algorithm is
[2]. Note that some mpi_* static functions have been made non static
and renamed mbedtls_mpi_* in the current Mbed TLS so the code from
v3.5.2 is modified accordingly.

Link: https://optee.readthedocs.io/en/latest/building/devices/qemu.html#qemu-v7 [1]
Link: https://github.com/Mbed-TLS/mbedtls/commit/1ba40585f9446372f69f19693c20f593283930af [2]
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>