ci: notify_maintainers: simplify scripts and use actions/github-script
Attempting once more to fix a 403 error in the notify_maintainers job.
This time, the notify_maintainers.py script does not post the comment
itself but just outputs the message it wants to post. It is then the
responsibility of the notify_maintainers job to post it, via the
actions/github-script action. In the tests I performed, the comment is
posted by "github-actions (bot)" and I received no 403 error.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>